remove composable, add intermediate component and correct file name
