+++ @xen-orchestra/web/src/components/vm/actions/VtsVmActions.vue
@@ -1,0 +1,12 @@
+<template>
+  <VmStartButton :vm />
+</template>
+
+<script lang="ts" setup>
+import type { XoVm } from '@vates/types'
+import VmStartButton from './VmStartButton.vue'
+
+const { vm } = defineProps<{
+  vm: XoVm
+}>()
+</script>
